externalize ObservationValidator
